lang: Use universal way to retrieve load
Sysinfo is not supported on macOS and results in a build error.
